flash-converter-ws/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flash-converter-ws"
dynamic = ["version"]
description = 'Web service to convert a video to mp3 quickly.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["video", "audio", "converter", "webservice"]
authors = [
    { name = "Laurent LAPORTE", email = "laurent.laporte.pro@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    # Project dependencies
    "flash-converter-tasks",
    # Third-party dependencies
    "fastapi",
    "uvicorn",
    "celery",
    "redis",
    "pydantic-settings",
    "python-multipart",
]

[project.urls]
Documentation = "https://github.com/laurent-laporte-pro/flash-converter-ws#readme"
Issues = "https://github.com/laurent-laporte-pro/flash-converter-ws/issues"
Source = "https://github.com/laurent-laporte-pro/flash-converter-ws"

[tool.hatch.version]
path = "src/flash_converter/ws/__about__.py"
attr = "__version__"

[tool.hatch.build.targets.wheel]
packages = ["src/flash_converter"]

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.12"]

# ----------------------------
# Type checking and linting
# ----------------------------

[tool.hatch.envs.types]
extra-dependencies = [
    "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --namespace-packages --package flash_converter.tasks --install-types --non-interactive {args:src/flash_converter/ws tests}"

# ----------------------------
# Configuration of ruff
# ----------------------------

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src", "tests"]
exclude = ["src/flash_converter/ws/__main__.py"]

# --------------------------------
# Configuration of tests coverage
# --------------------------------

[tool.coverage.run]
source_pkgs = ["flash_converter.ws", "tests"]
branch = true
parallel = true
omit = [
    "src/flash_converter/ws/__about__.py",
]

[tool.coverage.paths]
flash_converter_ws = ["src/flash_converter/ws", "*/flash-converter-ws/src/flash_converter/ws"]
tests = ["tests", "*/flash-converter-ws/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
